+++ adapter/fiberv3openapi/router_test.go
@@ -2,12 +2,10 @@ package fiberv3openapi_test
 
 import (
 	"io"
-	"mime/multipart"
 	"net/http"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/gofiber/fiber/v3"
 	"github.com/stretchr/testify/assert"
@@ -15,74 +13,12 @@ import (
 
 	stoplightemb "github.com/oaswrap/spec-ui/stoplightemb"
 	"github.com/oaswrap/spec/internal/testutil"
+	"github.com/oaswrap/spec/internal/testutil/dto"
 	"github.com/oaswrap/spec/openapi"
 	"github.com/oaswrap/spec/option"
 
 	"github.com/oaswrap/spec/adapter/fiberv3openapi"
 )
-
-type Pet struct {
-	ID        int      `json:"id"`
-	Name      string   `json:"name"`
-	Type      string   `json:"type"`
-	Status    string   `json:"status" enum:"available,pending,sold"`
-	Category  Category `json:"category"`
-	Tags      []Tag    `json:"tags"`
-	PhotoURLs []string `json:"photoUrls"`
-}
-
-type Tag struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-type Category struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-type UpdatePetWithFormRequest struct {
-	ID     int    `path:"petId" required:"true"`
-	Name   string `required:"true" formData:"name"`
-	Status string `formData:"status" enum:"available,pending,sold"`
-}
-
-type UploadImageRequest struct {
-	ID                 int64           `params:"petId" path:"petId"`
-	AdditionalMetaData string          `query:"additionalMetadata"`
-	_                  *multipart.File `contentType:"application/octet-stream"`
-}
-
-type DeletePetRequest struct {
-	ID     int    `path:"petId" required:"true"`
-	APIKey string `header:"api_key"`
-}
-
-type Order struct {
-	ID       int       `json:"id"`
-	PetID    int       `json:"petId"`
-	Quantity int       `json:"quantity"`
-	ShipDate time.Time `json:"shipDate"`
-	Status   string    `json:"status" enum:"placed,approved,delivered"`
-	Complete bool      `json:"complete"`
-}
-
-type PetUser struct {
-	ID         int    `json:"id"`
-	Username   string `json:"username"`
-	FirstName  string `json:"firstName"`
-	LastName   string `json:"lastName"`
-	Email      string `json:"email"`
-	Password   string `json:"password"`
-	Phone      string `json:"phone"`
-	UserStatus int    `json:"userStatus" enum:"0,1,2"`
-}
-
-type APIResponse struct {
-	Message string `json:"message"`
-	Type    string `json:"type"`
-	Code    int    `json:"code"`
-}
 
 func PingHandler(c fiber.Ctx) error {
 	return c.SendString("pong")
@@ -156,15 +92,15 @@ func TestRouter_Spec(t *testing.T) {
 					option.OperationID("updatePet"),
 					option.Summary("Update an existing pet"),
 					option.Description("Update the details of an existing pet in the store."),
-					option.Request(new(Pet)),
-					option.Response(200, new(Pet)),
+					option.Request(new(dto.Pet)),
+					option.Response(200, new(dto.Pet)),
 				)
 				pet.Post("/", nil).With(
 					option.OperationID("addPet"),
 					option.Summary("Add a new pet"),
 					option.Description("Add a new pet to the store."),
-					option.Request(new(Pet)),
-					option.Response(201, new(Pet)),
+					option.Request(new(dto.Pet)),
+					option.Response(201, new(dto.Pet)),
 				)
 				pet.Get("/findByStatus", nil).With(
 					option.OperationID("findPetsByStatus"),
@@ -173,7 +109,7 @@ func TestRouter_Spec(t *testing.T) {
 					option.Request(new(struct {
 						Status string `query:"status" enum:"available,pending,sold"`
 					})),
-					option.Response(200, new([]Pet)),
+					option.Response(200, new([]dto.Pet)),
 				)
 				pet.Get("/findByTags", nil).With(
 					option.OperationID("findPetsByTags"),
@@ -182,14 +118,14 @@ func TestRouter_Spec(t *testing.T) {
 					option.Request(new(struct {
 						Tags []string `query:"tags"`
 					})),
-					option.Response(200, new([]Pet)),
+					option.Response(200, new([]dto.Pet)),
 				)
 				pet.Post("/:petId/uploadImage", nil).With(
 					option.OperationID("uploadFile"),
 					option.Summary("Upload an image for a pet"),
 					option.Description("Uploads an image for a pet."),
-					option.Request(new(UploadImageRequest)),
-					option.Response(200, new(APIResponse)),
+					option.Request(new(dto.UploadImageRequest)),
+					option.Response(200, new(dto.APIResponse)),
 				)
 				pet.Get("/:petId", nil).With(
 					option.OperationID("getPetById"),
@@ -198,20 +134,20 @@ func TestRouter_Spec(t *testing.T) {
 					option.Request(new(struct {
 						ID int `uri:"petId" required:"true"`
 					})),
-					option.Response(200, new(Pet)),
+					option.Response(200, new(dto.Pet)),
 				)
 				pet.Post("/:petId", nil).With(
 					option.OperationID("updatePetWithForm"),
 					option.Summary("Update pet with form"),
 					option.Description("Updates a pet in the store with form data."),
-					option.Request(new(UpdatePetWithFormRequest)),
+					option.Request(new(dto.UpdatePetWithFormRequest)),
 					option.Response(200, nil),
 				)
 				pet.Delete("/{petId}", nil).With(
 					option.OperationID("deletePet"),
 					option.Summary("Delete a pet"),
 					option.Description("Delete a pet from the store by its ID."),
-					option.Request(new(DeletePetRequest)),
+					option.Request(new(dto.DeletePetRequest)),
 					option.Response(204, nil),
 				)
 				store := r.Group("/store").With(
@@ -221,8 +157,8 @@ func TestRouter_Spec(t *testing.T) {
 					option.OperationID("placeOrder"),
 					option.Summary("Place an order"),
 					option.Description("Place a new order for a pet."),
-					option.Request(new(Order)),
-					option.Response(201, new(Order)),
+					option.Request(new(dto.Order)),
+					option.Response(201, new(dto.Order)),
 				)
 				store.Get("/order/:orderId", nil).With(
 					option.OperationID("getOrderById"),
@@ -231,7 +167,7 @@ func TestRouter_Spec(t *testing.T) {
 					option.Request(new(struct {
 						ID int `uri:"orderId" required:"true"`
 					})),
-					option.Response(200, new(Order)),
+					option.Response(200, new(dto.Order)),
 					option.Response(404, nil),
 				)
 				store.Delete("/order/:orderId", nil).With(
@@ -251,15 +187,15 @@ func TestRouter_Spec(t *testing.T) {
 					option.OperationID("createUsersWithList"),
 					option.Summary("Create users with list"),
 					option.Description("Create multiple users in the store with a list."),
-					option.Request(new([]PetUser)),
+					option.Request(new([]dto.PetUser)),
 					option.Response(201, nil),
 				)
 				user.Post("/", nil).With(
 					option.OperationID("createUser"),
 					option.Summary("Create a new user"),
 					option.Description("Create a new user in the store."),
-					option.Request(new(PetUser)),
-					option.Response(201, new(PetUser)),
+					option.Request(new(dto.PetUser)),
+					option.Response(201, new(dto.PetUser)),
 				)
 				user.Get("/:username", nil).With(
 					option.OperationID("getUserByName"),
@@ -268,7 +204,7 @@ func TestRouter_Spec(t *testing.T) {
 					option.Request(new(struct {
 						Username string `uri:"username" required:"true"`
 					})),
-					option.Response(200, new(PetUser)),
+					option.Response(200, new(dto.PetUser)),
 					option.Response(404, nil),
 				)
 				user.Put("/:username", nil).With(
@@ -276,11 +212,11 @@ func TestRouter_Spec(t *testing.T) {
 					option.Summary("Update an existing user"),
 					option.Description("Update the details of an existing user."),
 					option.Request(new(struct {
-						PetUser
+						dto.PetUser
 
 						Username string `uri:"username" required:"true"`
 					})),
-					option.Response(200, new(PetUser)),
+					option.Response(200, new(dto.PetUser)),
 					option.Response(404, nil),
 				)
 				user.Delete("/:username", nil).With(